requirements.txt
vunit-hdl
pytest
tox
